+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## 1.17.0 - Unreleased
+- Add support for threaded abstract types. Threaded abstract types can easily be shared between threads.
 - Deprecate the `thread` library. Use threaded channels and ev instead.
 - Channels can now be marshalled.
 - Add the ability to close channels with `ev/chan-close` (or `:close`).